Updated interfaces acording to sqlx
