+++ compose.yml
@@ -1,9 +1,39 @@
 services:
   mysql:
-    image: mysql
-    restart: always
+    image: mysql:latest
     ports:
-      - 3306:3306
+      - 3313:3306
+    pull_policy: always
+    restart: always
+    environment:
+      MYSQL_ALLOW_EMPTY_PASSWORD: true
+      MYSQL_DATABASE: testdb
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "--user", "root"]
+      interval: 3s
+      timeout: 3s
+      retries: 10
+  mysql5:
+    image: mysql:5
+    platform: linux/amd64
+    ports:
+      - 3311:3306
+    pull_policy: always
+    restart: always
+    environment:
+      MYSQL_ALLOW_EMPTY_PASSWORD: true
+      MYSQL_DATABASE: testdb
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "--user", "root"]
+      interval: 3s
+      timeout: 3s
+      retries: 10
+  mysql8:
+    image: mysql:8
+    ports:
+      - 3312:3306
+    pull_policy: always
+    restart: always
     environment:
       MYSQL_ALLOW_EMPTY_PASSWORD: true
       MYSQL_DATABASE: testdb
@@ -13,10 +43,39 @@ services:
       timeout: 3s
       retries: 10
   mariadb:
-    image: mariadb
-    restart: always
+    image: mariadb:latest
     ports:
-      - 3307:3306
+      - 3316:3306
+    pull_policy: always
+    restart: always
+    environment:
+      MARIADB_ALLOW_EMPTY_ROOT_PASSWORD: true
+      MARIADB_DATABASE: testdb
+    healthcheck:
+      test: ["CMD", "mariadb-admin", "ping", "-h", "127.0.0.1"]
+      interval: 3s
+      timeout: 3s
+      retries: 10
+  mariadb10:
+    image: mariadb:10
+    ports:
+      - 3314:3306
+    pull_policy: always
+    restart: always
+    environment:
+      MARIADB_ALLOW_EMPTY_ROOT_PASSWORD: true
+      MARIADB_DATABASE: testdb
+    healthcheck:
+      test: ["CMD", "mariadb-admin", "ping", "-h", "127.0.0.1"]
+      interval: 3s
+      timeout: 3s
+      retries: 10
+  mariadb11:
+    image: mariadb:11
+    ports:
+      - 3315:3306
+    pull_policy: always
+    restart: always
     environment:
       MARIADB_ALLOW_EMPTY_ROOT_PASSWORD: true
       MARIADB_DATABASE: testdb
